+++ src/ClientLib/SnakeProject/settings.cpp
