+++ src/ClientLib/SnakeProject/settings.h
@@ -1,0 +1,11 @@
+#ifndef SNAKESETTINGS_H
+#define SNAKESETTINGS_H
+#include <quasarapp.h>
+
+#define THEME "THEME_GUI"
+
+#define THEME_DEFAULT 0
+
+using Settings = QuasarAppUtils::Settings;
+
+#endif // SNAKESETTINGS_H
