+++ src/ClientLib/SnakeProject/settingsviewmodel.h
@@ -2,14 +2,12 @@
 #define SETTINGSVIEWMODEL_H
 
 #include <QObject>
-#include "./../settings.h"
+#include "SnakeProject/settings.h"
 
 
 class SettingsViewModel: public QObject
 {
     Q_OBJECT
-    Q_PROPERTY(int port READ port WRITE setPort NOTIFY portChanged)
-    Q_PROPERTY(QString host READ host WRITE setHost NOTIFY hostChanged)
     Q_PROPERTY(int theme READ theme WRITE setTheme NOTIFY themeChanged)
 
 private:
@@ -20,8 +18,6 @@ private slots:
 
 public:
     SettingsViewModel(QObject* ptr = nullptr);
-    int port() const;
-    QString host() const;
     int theme() const;
 
     Q_INVOKABLE void forceUpdate();
@@ -29,13 +25,9 @@ public:
 
 
 public slots:
-    void setPort(int port);
-    void setHost(const QString &host);
     void setTheme(int theme);
 
 signals:
-    void portChanged(int port);
-    void hostChanged(const QString &host);
     void themeChanged(int theme);
 };
